requirements.txt
altgraph==0.17.3
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
Deprecated==1.2.14
epub==0.5.2
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
jaconv==0.3.4
Nuitka==2.3.11
ordered-set==4.1.0
outcome==1.2.0
pefile==2023.2.7
Pillow==9.5.0
pycparser==2.21
pyinstaller==5.12.0
pyinstaller-hooks-contrib==2023.3
pykakasi==2.2.1
PySocks==1.7.1
pywin32-ctypes==0.2.0
selenium==4.9.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.0
trio-websocket==0.10.2
typing-extensions==4.6.3
urllib3==2.0.3
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
zstandard==0.21.0
